pyproject.toml
[project]
name = "qtransformer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "marcel", email = "marcel.brunnbauer@gmx.at" }
]
requires-python = ">=3.10"
dependencies = [
    "hydra-core>=1.3.2",
    "numpy>=2.2.1",
    "python-dotenv>=1.0.1",
    "wandb>=0.19.1",
]
[project.optional-dependencies]
cpu = [
  "torch<2.3.0",
]
cu124 = [
  "torch<2.3.0",
]
d4rl = [
    "gym==0.23.1",
    "d4rl",
    "mujoco-py==2.1.2"
]

atari = [
    "gymnasium",
    "ale-py"
]

[tool.uv]
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu124" },
  ],
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu124", extra = "cu124" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
[project.scripts]
qtransformer = "qtransformer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
